Fix backcompat with OpenStack inventory
This moves the contents of the current dynamic inventory into a Python
module and adds two new inventories `inventory.py` and
`scaleup_inventory.py` that both use it.

`inventory.py` should work as before, introducing no breaking changes.

It does not work with the scaleup playbooks however, so those should use
the `scaleup_inventory.py` one.
